+++ app/components/MovieInputForm/DraftNewItem/index.tsx
@@ -1,13 +1,12 @@
 import { motion } from "motion/react";
 import type { DraftListItem } from "@/features/list/types/ListItem";
-import ListItemCard from "../../ListItem";
+import DraftListItemContainer from "../../ListItem/Draft";
 
 type Props = {
 	draft: DraftListItem;
-	isLoggedIn: boolean;
 };
 
-export default function DraftNewItem({ draft, isLoggedIn }: Props) {
+export default function DraftNewItem({ draft }: Props) {
 	return (
 		<motion.div
 			key="extracted-movie"
@@ -17,11 +16,7 @@ export default function DraftNewItem({ draft, isLoggedIn }: Props) {
 			transition={{ duration: 0.2, ease: "easeOut" }}
 			className="pt-4 px-4"
 		>
-			<ListItemCard
-				mode="extracted"
-				movie={draft}
-				isLoggedIn={isLoggedIn}
-			/>
+			<DraftListItemContainer draft={draft} />
 		</motion.div>
 	);
 }
