+++ features/auth/state/authAtom.ts
@@ -1,0 +1,16 @@
+"use client";
+
+import { useAtomValue } from "jotai";
+import { atom } from "jotai";
+
+export const authAtom = atom<{
+	isLoggedIn: boolean;
+	publicListId: string | null;
+}>({
+	isLoggedIn: false,
+	publicListId: null,
+});
+
+export const useAuth = () => {
+	return useAtomValue(authAtom);
+};
